+++ .hooks/pre-commit
@@ -13,3 +13,12 @@ if [ $vetcount -gt 0 ]; then
     exit 1
 fi
 exit 0
+
+# Ensure FIXME lines are removed before commit.
+fixme_lines=$(git diff --cached | grep ^+ | grep -v pre-commit | grep FIXME | sed 's_^+\s*__g')
+if [ "$todo_lines" != "" ]; then
+    echo "Please remove the following lines:"
+    echo -e "$todo_lines"
+    exit 1
+fi
+
